+++ src/components/auth/InputField.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { TextInput } from 'react-native';
 import { KeyboardTypeOptions, StyleSheet } from 'react-native';
-import Colors from '../../assets/Colors';
+import Colors from '../../../assets/Colors';
 
 type InputFieldProps = {
   onChange: (text: string) => void;
