+++ src/screens/auth/LoginScreen.tsx
@@ -5,15 +5,15 @@ import {
   Row,
   SafeArea,
   ButtonContainer,
-} from '../../common/Containers';
+} from '../../components/common/Containers';
 import {
   BodySubtext,
   BodyText,
   CenterText,
   H1Heading,
 } from '../../../assets/Fonts';
-import { BackButton, PurpleButton } from '../../common/Buttons';
-import InputField from '../../common/InputField';
+import { BackButton, PurpleButton } from '../../components/common/Buttons';
+import InputField from '../../components/auth/InputField';
 import { styles } from './styles';
 
 export default function LoginScreen() {
