+++ src/screens/auth/SignUpScreen.tsx
@@ -4,10 +4,10 @@ import {
   HeadingContainer,
   SafeArea,
   ButtonContainer,
-} from '../../common/Containers';
+} from '../../components/common/Containers';
 import { BodyText, H1Heading } from '../../../assets/Fonts';
-import { BackButton, PurpleButton } from '../../common/Buttons';
-import InputField from '../../common/InputField';
+import { BackButton, PurpleButton } from '../../components/common/Buttons';
+import InputField from '../../components/auth/InputField';
 
 export default function SignUpScreen() {
   const [email, setEmail] = useState('');
